01-azure-rag/app/requirements.txt
# requirements.txt - Azure RAG Application
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
psycopg2-binary==2.9.9
pgvector==0.2.4
python-multipart==0.0.6
pydantic==2.6.0
